+++ app/components/navigation_bar.html.erb
