+++ app/components/notification_banner.html.erb
