+++ app/views/layouts/component_preview.html.erb
@@ -24,7 +24,7 @@
   <body class="govuk-template__body">
     <%= render "layouts/add_js_enabled_class_to_body" %>
 
-    <%= render Find::Header::View.new(
+    <%= render Find::Header.new(
       service_name: "Find and Publish Components",
     ) %>
