+++ app/views/layouts/publish.html.erb
@@ -1,6 +1,6 @@
 <%= extends_layout :application do %>
   <%= content_for(:header) do %>
-    <%= render Header::View.new(
+    <%= render Header.new(
       service_name: I18n.t("service_name.publish"),
       current_user:,
     ) %>
