+++ app/views/layouts/support.html.erb
@@ -1,6 +1,6 @@
 <%= extends_layout :application do %>
   <%= content_for(:header) do %>
-    <%= render Header::View.new(
+    <%= render Header.new(
       service_name: I18n.t("service_name.support"),
       current_user:,
     ) %>
@@ -8,7 +8,7 @@
 
 <!-- Uncomment this when more items are needed on the navigation bar.
   <%# <%= content_for(:navigation_bar) do %> %>
-    <%# <%= render NavigationBar::View.new( %>
+    <%# <%= render NavigationBar.new( %>
       <%# items: [ %>
         <%# { name: "Providers", url: support_providers_path, current: true }, %>
       <%# ], %>
