+++ app/views/publish/shared/_navigation_bar.html.erb
@@ -1,5 +1,5 @@
 <% if render_navigation_bar?(@provider) %>
-  <%= render(NavigationBar::View.new(
+  <%= render(NavigationBar.new(
     items: navigation_items(@provider),
     current_path: request.path,
     current_user: @current_user,
