+++ spec/components/find_interface/header/view_preview.rb
@@ -3,11 +3,11 @@
 module Find::Header
   class ViewPreview < ViewComponent::Preview
     def with_custom_service_name
-      render(Header::View.new(service_name: "Hello"))
+      render(Header.new(service_name: "Hello"))
     end
 
     def with_our_service_name
-      render(Header::View.new(service_name: I18n.t("service_name.find")))
+      render(Header.new(service_name: I18n.t("service_name.find")))
     end
   end
 end
